% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.function.R
\name{example.function}
\alias{example.function}
\title{example.function}
\usage{
example.function(x, y)
}
\arguments{
\item{x}{Numeric scalar.}

\item{y}{Numeric scalar.}
}
\value{
Numeric scalar. The sum of \code{x} and \code{y}.
}
\description{
Compute the sum of two numeric inputs. This function expects numeric
scalars for both arguments and returns their sum.
}
\examples{
example.function(
    x = 1,
    y = 2
    );
}
