% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatPlot-package.R
\docType{package}
\name{StatPlot-package}
\alias{StatPlot}
\alias{StatPlot-package}
\title{StatPlot: Interactive Data Visualization with ggplot2}
\description{
Shiny application for interactive data visualization using ggplot2. Allows users to create, customize and combine charts without programming knowledge. Part of the StatSuite ecosystem developed at ICOMVIS, Universidad Nacional, Costa Rica.
}
\author{
\strong{Maintainer}: Manuel Spínola \email{manuel.spinola@una.ac.cr}

Authors:
\itemize{
  \item Manuel Spínola \email{manuel.spinola@una.ac.cr}
}

}
\keyword{internal}
