% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_pvalues.R
\name{combine_pvalues}
\alias{combine_pvalues}
\title{Combine p-values}
\usage{
combine_pvalues(p)
}
\arguments{
\item{p}{Numeric vector of p-values.}
}
\value{
A named vector of combined p-values for each method
}
\description{
Combines a numeric vector of p-values into a single p-value using a variety of methods.  Note CMC or MCM are recommended by Chen 2022.
}
\examples{
pvals <- c(0.001, 0.005, 0.2)
combine_pvalues(pvals)
}
\references{
Chen, Z. Robust tests for combining p-values under arbitrary dependency structures. Sci Rep 12, 3158 (2022). https://doi.org/10.1038/s41598-022-07094-7
}
