% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_overall.R
\name{table_overall}
\alias{table_overall}
\title{Create a table that summarizes the entire cohort.}
\usage{
table_overall(d)
}
\arguments{
\item{d}{data.frame Dataset containing variables to summarize.}
}
\value{
A gtsummary::tbl_summary object for the overall cohort.
}
\description{
Convenience wrapper that builds a gtsummary::tbl_summary for the overall
dataset.
}
\examples{
df <- data.frame(age = rnorm(20, 50, 10), sex = sample(c("M","F"), 20, TRUE))
table_overall(df)
}
