% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_bw2}
\alias{theme_bw2}
\title{Bold-styled Black and White Theme}
\usage{
theme_bw2()
}
\value{
A ggplot2 theme object
}
\description{
A modified ggplot2 black and white theme with bold text elements.
}
\examples{
ggplot(mtcars, aes(x = wt, y = mpg)) +
  geom_point() +
  theme_bw2()
}
