% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_classic2}
\alias{theme_classic2}
\title{Bold-styled Classic Theme}
\usage{
theme_classic2()
}
\value{
A ggplot2 theme object
}
\description{
A modified ggplot2 classic theme with bold text elements and centered title.
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(x = wt, y = mpg)) +
  geom_point() +
  theme_classic2()
}
